pyproject.toml
[build-system]
requires = ["hatchling>=1.20"]
build-backend = "hatchling.build"

[project]
name = "procgen-textures"
version = "0.1.0"
description = "Procedural 2D tileable textures (albedo/height/normal) for game dev prototypes — Python CLI with NumPy & Pillow."
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [{ name = "Cem (gebirgshang)" }]
keywords = ["procedural", "textures", "tileable", "perlin", "worley", "gamedev", "normal-map"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "License :: OSI Approved :: MIT License",
  "Topic :: Multimedia :: Graphics",
  "Topic :: Games/Entertainment",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent"
]
dependencies = [
  "numpy>=1.24",
  "Pillow>=10.0"
]

[project.urls]
Homepage = "https://github.com/USER/REPO"
Issues = "https://github.com/USER/REPO/issues"

[project.scripts]
procgen-textures = "procgen_textures.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/procgen_textures"]
